requirements.txt
# Dashboard
streamlit>=1.25.0
 pyyaml

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data sources
requests>=2.31.0

# Already have:
numpy>=1.19.1
pandas>=1.1.1
yfinance>=0.1.54
scipy>=1.5.2
matplotlib>=3.3.1